# Conflicts:
#	src/main/java/sim/danslchamp/circuit/Circuit.java
#	src/main/java/sim/danslchamp/circuit/Résistor.java

+++ src/main/java/sim/danslchamp/circuit/Fil.java
@@ -34,7 +34,7 @@ public class Fil extends Composant {
     }
 
     @Override
-    public double calculResistance(int frequence) {
+    public double calculResistance(double frequence) {
         return 0;
     }
